requirements.txt
pymongo
opencv-python
numpy
pillow
opencv-contrib-python
tqdm
python-dotenv